Isolated current problems in long-long code
When calling a function that takes a long-long as an argument improper registers are used.
For example:
In this call to `printf()` with a long-long argument the following code is generated
```
lw      a2,-56(s0)
lw      a0,-52(s0)
auipc   a0,0x0
mv      a0,a0
auipc   ra,0x0
jalr    ra # 1d0 <main+0x1d0>
```
a1 should be used instead of a2 (I think)

+++ riscv_utils.h
@@ -33,15 +33,16 @@ void emit_J( uint32_t imm, uint32_t rd, uint32_t opcode );
 
 // Macros for masking values for immediate operations
 // mask off the lower 12 bits of a 32-bit value
-#define IMM_LOW( imm ) ( (imm)&0x00000FFF )
+#define IMM_LOW( imm ) ( ( imm ) & 0x00000FFF )
 // mask and shift a 32-bit immediate value to grab the upper 24 bits
-#define IMM_HIGH( imm ) ( ( (imm)&0xFFFFF000 ) >> 12 )
+#define IMM_HIGH( imm ) ( ( ( imm ) & 0xFFFFF000 ) >> 12 )
 // like IMM_HIGH but with lower sign extended bits considered
 // so when the lower (sign extended) bits get added, they don't ruin things
-#define IMM_HIGH_LEXT( imm ) ( IMM_HIGH( (imm) ) + (((imm) >> 11) & 1) )
+#define IMM_HIGH_LEXT( imm ) ( IMM_HIGH( ( imm ) ) + ( ( ( imm ) >> 11 ) & 1 ) )
 
 // macro to check if the immediate value will be larger than 12 bits
-#define LARGE_IMM( imm ) ( (uint32_t)( IMM_HIGH( ((unsigned long long)imm + (1 << 11)) ) ) )
+#define LARGE_IMM( imm ) ( (uint32_t)( IMM_HIGH( ( (uint32_t)( imm ) + ( 1 << 11 ) ) ) ) )
+// #define LARGE_IMM( imm ) ( (uint32_t)( IMM_HIGH( ( imm ) + 0x800 ) ) )
 
 // Now for a big table of opcodes (RV32I) from p130 of ISA documentation
 // https://github.com/riscv/riscv-isa-manual/releases/download/Ratified-IMAFDQC/riscv-spec-20191213.pdf
@@ -137,7 +138,7 @@ void emit_J( uint32_t imm, uint32_t rd, uint32_t opcode );
 #define emit_RET() ( emit_JALR( 0, 1, 0 ) )
 
 #define emit_MV( rd, rs ) ( emit_ADDI( rd, rs, 0 ) )
-//#define emit_SEXT_W(rd, rs) (emit_ADDIW(rd, rs, 0))
+// #define emit_SEXT_W(rd, rs) (emit_ADDIW(rd, rs, 0))
 
 // branches
 // clang-format off
